Move couch active tasks service to libs
