+++ src/commands/doc/replicate.ts
@@ -2,7 +2,7 @@ import { Args, Command, Options } from '@effect/cli';
 import { Array, Console, Effect, Option, pipe, Predicate, Stream } from 'effect';
 import { initializeUrl } from '../../index';
 import { ReplicateService, ReplicationDoc } from '../../services/replicate';
-import { CouchActiveTask, streamActiveTasks } from '../../services/couch/active-tasks';
+import { CouchActiveTask, streamActiveTasks } from '../../libs/couch/active-tasks';
 import { ParseError } from 'effect/Cron';
 import { clearThen } from '../../libs/console';
